feat: commenting perms

+++ drive/api/files.py
@@ -988,6 +988,8 @@ def create_comment(parent, name, content, is_reply):
 @frappe.whitelist()
 def edit_comment(name, content):
     comment = frappe.get_doc("Drive Comment", name)
+    if comment.owner != frappe.session.user:
+        raise PermissionError("You can't edit comments you don't own.")
     comment.content = content
     comment.save()
     return name
@@ -996,6 +998,8 @@ def edit_comment(name, content):
 @frappe.whitelist()
 def delete_comment(name, entire=True):
     comment = frappe.get_doc("Drive Comment", name)
+    if comment.owner != frappe.session.user:
+        raise PermissionError("You can't edit comments you don't own.")
     if entire:
         for r in comment.replies:
             r.delete()
